# Add arena helper functions and improve block insertion