This PR introduces several helper functions into the arena allocator and refactors the block insertion strategy in `arena_alloc` to improve performance. As a result, total allocation times have been measurably reduced.

Below is the benchmark script and corresponding results demonstrating the speedup:

## Hardware info

```
CPU: AMD Ryzen 5 7535HS with Radeon Graphics
Cores/Threads: 12 threads
RAM: 30Gi
OS: Ubuntu 24.04.2 LTS
Kernel: 6.11.0-26-generic
Compiler: gcc (Ubuntu 13.3.0-6ubuntu2~24.04) 13.3.0
```

## Benchmark Script on Stage-0

<details>
<summary>Benchmark Script</summary>

```sh
#!/bin/bash
export LC_ALL=C

n=15
warmup=5

total_user=0
total_sys=0
total_elapsed=0
total_rss=0

tmp_file="$(mktemp)"

echo "Warming up $warmup times..."
for i in $(seq 1 $warmup); do
    ./out/shecc ./src/main.c >/dev/null 2>&1
done

echo "Running $n benchmarks..."
for i in $(seq 1 $n); do
    /usr/bin/time -f "%U %S %e %M" -o "$tmp_file" ./out/shecc ./src/main.c >/dev/null 2>&1

    read user sys elapsed rss < "$tmp_file"

    echo "Run $i: user=${user}s sys=${sys}s elapsed=${elapsed}s maxrss=${rss}KB"

    total_user=$(awk "BEGIN {printf \"%.6f\", $total_user + $user}")
    total_sys=$(awk "BEGIN {printf \"%.6f\", $total_sys + $sys}")
    total_elapsed=$(awk "BEGIN {printf \"%.6f\", $total_elapsed + $elapsed}")
    total_rss=$(awk "BEGIN {print $total_rss + $rss}")
done

rm -f "$tmp_file"

avg_user=$(awk "BEGIN {printf \"%.6f\", $total_user / $n}")
avg_sys=$(awk "BEGIN {printf \"%.6f\", $total_sys / $n}")
avg_elapsed=$(awk "BEGIN {printf \"%.6f\", $total_elapsed / $n}")
avg_rss=$(awk "BEGIN {printf \"%.2f\", $total_rss / $n}")

echo "----------------------------------"
echo "Average user time:    ${avg_user}s"
echo "Average system time:  ${avg_sys}s"
echo "Average elapsed time: ${avg_elapsed}s"
echo "Average max RSS:      ${avg_rss} KB"
```
</details>

## Before (Stage-0)

<details>
<summary>Result</summary>

```
Warming up 5 times...
Running 15 benchmarks...
Run 1: user=0.16s sys=0.15s elapsed=0.32s maxrss=291848KB
Run 2: user=0.16s sys=0.15s elapsed=0.32s maxrss=291528KB
Run 3: user=0.14s sys=0.16s elapsed=0.30s maxrss=291656KB
Run 4: user=0.14s sys=0.16s elapsed=0.31s maxrss=291912KB
Run 5: user=0.16s sys=0.16s elapsed=0.32s maxrss=291912KB
Run 6: user=0.15s sys=0.17s elapsed=0.33s maxrss=291652KB
Run 7: user=0.15s sys=0.16s elapsed=0.32s maxrss=291848KB
Run 8: user=0.16s sys=0.15s elapsed=0.32s maxrss=291720KB
Run 9: user=0.15s sys=0.15s elapsed=0.31s maxrss=291524KB
Run 10: user=0.15s sys=0.16s elapsed=0.31s maxrss=291848KB
Run 11: user=0.15s sys=0.16s elapsed=0.32s maxrss=291720KB
Run 12: user=0.15s sys=0.15s elapsed=0.31s maxrss=291912KB
Run 13: user=0.16s sys=0.14s elapsed=0.31s maxrss=291720KB
Run 14: user=0.14s sys=0.16s elapsed=0.31s maxrss=291784KB
Run 15: user=0.16s sys=0.14s elapsed=0.31s maxrss=291656KB
----------------------------------
Average user time:    0.152000s
Average system time:  0.154667s
Average elapsed time: 0.314667s
Average max RSS:      291749.33 KB
```
</details>

## After (Stage-0)

<details>
<summary>Result</summary>

```
Warming up 5 times...
Running 15 benchmarks...
Run 1: user=0.08s sys=0.15s elapsed=0.23s maxrss=294344KB
Run 2: user=0.09s sys=0.14s elapsed=0.24s maxrss=294344KB
Run 3: user=0.06s sys=0.16s elapsed=0.22s maxrss=294408KB
Run 4: user=0.06s sys=0.15s elapsed=0.22s maxrss=294536KB
Run 5: user=0.06s sys=0.16s elapsed=0.22s maxrss=294344KB
Run 6: user=0.06s sys=0.15s elapsed=0.22s maxrss=294344KB
Run 7: user=0.05s sys=0.16s elapsed=0.22s maxrss=294472KB
Run 8: user=0.07s sys=0.15s elapsed=0.22s maxrss=294408KB
Run 9: user=0.06s sys=0.16s elapsed=0.23s maxrss=294408KB
Run 10: user=0.06s sys=0.15s elapsed=0.22s maxrss=294600KB
Run 11: user=0.06s sys=0.15s elapsed=0.22s maxrss=294472KB
Run 12: user=0.07s sys=0.17s elapsed=0.25s maxrss=294088KB
Run 13: user=0.07s sys=0.15s elapsed=0.23s maxrss=294024KB
Run 14: user=0.08s sys=0.14s elapsed=0.23s maxrss=294088KB
Run 15: user=0.06s sys=0.15s elapsed=0.21s maxrss=294536KB
----------------------------------
Average user time:    0.066000s
Average system time:  0.152667s
Average elapsed time: 0.225333s
Average max RSS:      294361.07 KB
``` 
</details>

## Benchmark Script on Stage-1

<details>
<summary>Benchmark Script</summary>

```bash
#!/bin/bash
export LC_ALL=C

n=15
warmup=5

total_user=0
total_sys=0
total_elapsed=0
total_rss=0

tmp_file="$(mktemp)"

echo "Warming up $warmup times..."
for i in $(seq 1 $warmup); do
    qemu-arm ./out/shecc-stage1.elf ./src/main.c >/dev/null 2>&1
done

echo "Running $n benchmarks..."
for i in $(seq 1 $n); do
    /usr/bin/time -f "%U %S %e %M" -o "$tmp_file" qemu-arm ./out/shecc-stage1.elf ./src/main.c >/dev/null 2>&1

    read user sys elapsed rss < "$tmp_file"

    echo "Run $i: user=${user}s sys=${sys}s elapsed=${elapsed}s maxrss=${rss}KB"

    total_user=$(awk "BEGIN {printf \"%.6f\", $total_user + $user}")
    total_sys=$(awk "BEGIN {printf \"%.6f\", $total_sys + $sys}")
    total_elapsed=$(awk "BEGIN {printf \"%.6f\", $total_elapsed + $elapsed}")
    total_rss=$(awk "BEGIN {print $total_rss + $rss}")
done

rm -f "$tmp_file"

avg_user=$(awk "BEGIN {printf \"%.6f\", $total_user / $n}")
avg_sys=$(awk "BEGIN {printf \"%.6f\", $total_sys / $n}")
avg_elapsed=$(awk "BEGIN {printf \"%.6f\", $total_elapsed / $n}")
avg_rss=$(awk "BEGIN {printf \"%.2f\", $total_rss / $n}")

echo "----------------------------------"
echo "Average user time:    ${avg_user}s"
echo "Average system time:  ${avg_sys}s"
echo "Average elapsed time: ${avg_elapsed}s"
echo "Average max RSS:      ${avg_rss} KB"
```
</details>

## Before (Stage-1)

<details>
<summary>Result</summary>

```
Warming up 5 times...
Running 15 benchmarks...
Run 1: user=1.36s sys=1.12s elapsed=2.49s maxrss=638864KB
Run 2: user=1.35s sys=1.14s elapsed=2.50s maxrss=639104KB
Run 3: user=1.38s sys=1.10s elapsed=2.49s maxrss=638936KB
Run 4: user=1.35s sys=1.14s elapsed=2.49s maxrss=638864KB
Run 5: user=1.36s sys=1.10s elapsed=2.47s maxrss=638664KB
Run 6: user=1.36s sys=1.09s elapsed=2.46s maxrss=638884KB
Run 7: user=1.34s sys=1.10s elapsed=2.46s maxrss=638920KB
Run 8: user=1.36s sys=1.10s elapsed=2.47s maxrss=638796KB
Run 9: user=1.39s sys=1.10s elapsed=2.50s maxrss=638600KB
Run 10: user=1.34s sys=1.10s elapsed=2.45s maxrss=638800KB
Run 11: user=1.40s sys=1.08s elapsed=2.49s maxrss=638684KB
Run 12: user=1.36s sys=1.09s elapsed=2.47s maxrss=639060KB
Run 13: user=1.41s sys=1.07s elapsed=2.49s maxrss=638712KB
Run 14: user=1.36s sys=1.12s elapsed=2.49s maxrss=639016KB
Run 15: user=1.35s sys=1.10s elapsed=2.48s maxrss=638500KB
----------------------------------
Average user time:    1.364667s
Average system time:  1.103333s
Average elapsed time: 2.480000s
Average max RSS:      638826.93 KB
```
</details>

## After (Stage-1)

<details>
<summary>Result</summary>

```
Warming up 5 times...
Running 15 benchmarks...
Run 1: user=0.73s sys=1.03s elapsed=1.77s maxrss=435132KB
Run 2: user=0.72s sys=1.06s elapsed=1.80s maxrss=434956KB
Run 3: user=0.70s sys=1.05s elapsed=1.76s maxrss=435108KB
Run 4: user=0.69s sys=1.05s elapsed=1.76s maxrss=435104KB
Run 5: user=0.67s sys=1.06s elapsed=1.74s maxrss=435448KB
Run 6: user=0.69s sys=1.04s elapsed=1.74s maxrss=435168KB
Run 7: user=0.68s sys=1.05s elapsed=1.75s maxrss=435396KB
Run 8: user=0.68s sys=1.06s elapsed=1.75s maxrss=435556KB
Run 9: user=0.69s sys=1.09s elapsed=1.79s maxrss=434892KB
Run 10: user=0.69s sys=1.05s elapsed=1.74s maxrss=435560KB
Run 11: user=0.69s sys=1.06s elapsed=1.76s maxrss=434988KB
Run 12: user=0.72s sys=1.02s elapsed=1.75s maxrss=434968KB
Run 13: user=0.68s sys=1.05s elapsed=1.75s maxrss=435432KB
Run 14: user=0.69s sys=1.05s elapsed=1.76s maxrss=434868KB
Run 15: user=0.71s sys=1.04s elapsed=1.76s maxrss=435092KB
----------------------------------
Average user time:    0.695333s
Average system time:  1.050667s
Average elapsed time: 1.758667s
Average max RSS:      435177.87 KB
```
</details>

 <div id='description'>
    <a href="https://bito.ai#summarystart"></a>
<h3>Summary by Bito</h3>
This pull request enhances arena memory management by adding new helper functions and refactoring the block insertion strategy in `arena_alloc`, resulting in improved performance and reduced allocation times. Error handling has also been enhanced with clearer failure messages, increasing code robustness.
<!-- Disabling unit_tests and post_effort_to_review
<br>
<br>
<b>Unit tests added</b>: False
<br>
<br>
<b>Estimated effort to review (1-5, lower is better)</b>: 2
-->
</div>