@AW-AlanWu 2025-06-01
Based on my current observations, most of the memory allocation in shecc follow a unified allocate-and-release model. This makes almost all scenarios suitable for replacing the current use of malloc, calloc, and free with an arena allocator. I plan to start working on this replacement soon.

In the short term, I'm also working on improving the current `strbuf_t` implementation by introducing a general-purpose dynamic array in shecc. This new array structure can not only replace the existing `strbuf_t`, but also serve as a replacement for many of the fixed-size arrays currently used in shecc. In scenarios with frequent push operations, using realloc can lead to performance improvements. 

Additionally, I’m exploring optimizations like improving cache locality by replacing the linked list used in the internal structure of the hashmap.